requirements.txt
langchain
python-dotenv
ipykernel
langchain_groq
langchain_core
langserve
uvicorn
fastapi
sse_starlette